Adding a run all command and also descriptions and tests for problems 6-10 (tests will fail)

+++ ProjectEuler/ProjectEuler/Problems.cs
@@ -177,7 +177,7 @@ public class Problems
         return 0;
     }
 
-    public static long Problem9()
+    public static int Problem9()
     {
         Console.WriteLine(ProblemSummaries.Problem9);
